cleaning up books components

+++ routes/api.js
@@ -5,15 +5,38 @@ const { Types } = require('mongoose');
 // Get all users or a specific user by id.
 router.get('/users', (req, res) => {
     console.log(req.path);
-    const id = req.body ? req.body.id : null;
+    if (!req.body) res.sendStatus(400);
+    const { id } = req.body;
     const promise = id ? User.findById(Types.ObjectId(id)) : User.find();
     promise.then(res.send).catch(res.send);
 });
 
-// Delete a user by id.
+const deepDelete = (from, to) => {
+    for (let item in from) {
+        const value = from[item];
+        if (typeof value === 'object')
+            deepDelete(value, to[item]);
+        else
+            delete to[item];
+    }
+};
+
+// Delete a user by id or delete fields from user as specified in body.
 router.delete('/users', (req, res) => {
-    const id = req.body ? req.body.id : null;
-    User.findByIdAndDelete(Types.ObjectId(id)).then(res.send).catch(res.send);
+    if (!req.body) res.sendStatus(400);
+    const { id } = req.body.id;
+    if (!id) res.sendStatus(400);
+    delete req.body.id;
+    if (req.body.keys)
+        // Delete fields from the user.
+        User.findByIdAndUpdate(Types.ObjectId(id)).then(user => {
+            deepDelete(req.body, user);
+            res.send(user);
+        }).catch(res.send);
+    else
+        // Delete the user.
+        User.findByIdAndDelete(Types.ObjectId(id)).then(res.send).catch(res.send);
+
 });
 
 // Create a new user.
@@ -24,8 +47,8 @@ router.post('/users', (req, res) => {
 // Get all movies, a specific movie by id, or titles starting with given title string.
 router.get('/movies', (req, res) => {
     console.log(req.path);
-    const id = req.body ? req.body.id : null;
-    const title = req.body ? req.body.title : null;
+    if (!req.body) res.sendStatus(400);
+    const { id, title } = req.body;
     const promise = id ? Movie.findById(Types.ObjectId(id)) :
         title ? Movie.find({ title: new RegEx('^' + title) }) :
             Movie.find();
@@ -34,7 +57,8 @@ router.get('/movies', (req, res) => {
 
 // Delete a movie by id.
 router.delete('/movies', (req, res) => {
-    const id = req.body ? req.body.id : null;
+    if (!req.body) res.sendStatus(400);
+    const { id } = req.body;
     movie.findByIdAndDelete(Types.ObjectId(id)).then(res.send).catch(res.send);
 });
 
@@ -43,21 +67,18 @@ router.post('/movies', (req, res) => {
     movie.create(req.body).then(res.send).catch(res.send);
 });
 
+// Copy all members/elements in object from into object to, converting ids and movies into ObjectIds.
 const deepCopy = (from, to) => {
-    const copy = (value, item) => {
-        if (typeof value == 'object')
+    for (let item in from) {
+        const value = from[item];
+        if (typeof value === 'object')
             deepCopy(value, to[item]);
         else
             to[item] = item === 'id' || item === 'movie' ? Types.ObjectId(value) : value;
-    };
-    if (Array.isArray(from))
-        from.forEach(copy);
-    else
-        for (let item in from)
-            copy(from[item], item);
+    }
 };
 
-// Modify a user with given id. All fields in body will be copied into given user.
+// Modify a user with given id. All fields in body will be copied into user.
 router.put('/users', (req, res) => {
     if (!req.body) res.sendStatus(400);
     const { id } = req.body;
